[fix] Support local mode without setting CUDA_VISIBLE_DEVICES (#38)
If `CUDA_VISIBLE_DEVICES` is not set, it means that all the
devices will be available for use. If it's an empty string,
it means no device is available for the process to use.

As a result, if `CUDA_VISIBLE_DEVICES` does not exist, the
default value should be equivalent to all the devices instead
of an empty one.

Signed-off-by: Hollow Man <hollowman@opensuse.org>